[HER-40]Change path
