+++ src/core/layout/footer/content.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link, Typography } from '@material-ui/core'
 
-import { useStyles } from 'core/footer/styles'
+import { useStyles } from 'core/layout/footer/styles'
 
 export const Content = () => <Typography className={useStyles().title}>
   Made with &#x2764; by
