+++ src/core/layout/footer/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Box } from '@material-ui/core'
 
-import { Content } from 'core/footer/content'
-import { useStyles } from 'core/footer/styles'
+import { Content } from 'core/layout/footer/content'
+import { useStyles } from 'core/layout/footer/styles'
 
 export const Footer = () => <Box className={useStyles().footer} display={'flex'} justifyContent={'center'}
                                  alignItems={'center'}>
