+++ src/core/layout/navbar/appbar.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { AppBar as MaterialAppBar, Grid } from '@material-ui/core'
 import PropTypes from 'prop-types'
 
-import { useStyles } from 'core/navbar/style'
+import { useStyles } from 'core/layout/navbar/style'
 
 export const AppBar = ({ children }) => <MaterialAppBar position={'static'}>
   <Grid container className={useStyles().root} alignItems={'center'} justify={'space-between'}>
