+++ src/core/layout/navbar/content/email.jsx
