+++ src/core/layout/navbar/content/gravatar.jsx
