--- src/core/navbar/index.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-import { Title } from 'core/navbar/title'
-import { UserContent } from 'core/navbar/content'
-import { AppBar } from 'core/navbar/appbar'
-
-export const Navbar = () => <AppBar>
-  <Title />
-  <UserContent />
-</AppBar>
